relocate helper functions
